requirement.txt
pandas
duckdb
psycopg2-binary
sqlalchemy
python-dotenv
faker
